# libobs: Remove unnecessary frame pipelining