@palana 2019-05-22
What are the timings like for other graphics cards, and how do timings change when the GPU is simultaneously being used for other things (such as games)? I'd assume pipelining to make a difference while there's other work going on, but not sure if anyone ever actually tested the impact