@jpark37 2019-05-22
I did run Intel GPA frame timing on my NVIDIA RTX 2080 Ti, and the numbers were about the same.

Let me try to diagram this, top path is new, bottom path is old:
![image](https://user-images.githubusercontent.com/10396506/58201636-672f6080-7c8a-11e9-89d4-e8ea0a3aaf12.png)

The key stall we want to avoid is the CPU waiting on the GPU to download the results of D. For example, if you tell D3D, "draw D, and then map D for read" then you will get a lengthy stall waiting for the GPU to catch up. However, if you say "draw current D, and then map previous D for read" the GPU is very likely to be done with D from the previous frame. There's just no reason to double-buffer A, B, or C.

Okay, I lied, there may be a reason. You might notice that A1/BO, B1/C0, and C1/D0 have no dependencies on each other, so maybe a GPU could schedule them more efficiently in theory. Switching a render target immediately from write to read as the PR does can introduce resource barrier penalties. In practice, I am not seeing this potential overlap produce a win, and as mentioned in the change description, Intel has worse cache behavior.

This diagram also doesn't go over the NVENC path, but the reasoning is similar.