@mufunyo 2019-05-25
If you put optimisations like this in the sRGB PR instead of a separate PR, you can claim it doesn't have a negative performance impact. :p