@jpark37 2019-05-26
IIRC, It's not quite enough savings to pay for the sRGB format perf loss on Intel Gen9. I have some other ideas for potential format conversion speed-ups like incorporating MRT and simplifying shaders, which may be desirable anyway to address mantissa precision loss and artifacts at high resolutions. Compute shaders would be ideal, but the current baseline doesn't require them. For one, macOS requires Metal for CS support, and it sounds like we won't drop GL support when that gets implemented.

Also, I'm finding GPU profiling on PC is a bit of a nightmare. D3D/GL timings don't match tools, tools don't match each other, copy timings are especially weird, and all timing sources report numbers with huge variances. I try to compare minimum timings in general, but sometimes the numbers I get in some situations are suspiciously low.