@h1z1 2019-05-26
Are those timings based on average, min, max and over how many frames ?  I'd expect higher bandwidth use since your'e transfering more it's kinda the point.