@jpark37 2019-05-26
Min of several GPA refreshes, which itself is hopefully not just playing back timings from a single run.

Why would bandwidth usage increase with this PR? It should be lower since cache utilization is higher.