@h1z1 2019-05-27
Swear I replied to this, hm.   I meant the system bandwidth not stream and go down with the change not up.  I'd expect however (as in consider OK),  higher bandwidth use with more frames.  Different processors of course have different cache sizes, the net effect is not universal.   Is there something I'm missing as to why this wouldn't be configurable vs removed entirely?