@jpark37 2019-05-27
It's not configurable because I don't see a benefit to the current render target setup. This PR doesn't change the number of draws, or even the number of pixels involved in those draws. It's doing the exact same amount of work, just better.